requirements.txt
Django==3.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
